# backend/cmd/app/docs/swagger.yaml
basePath: /
definitions:
...
  normaladmin_backend_internal_models.Menu:
    properties:
      api_method:
        description: 新增字段
        type: string
      api_path:
        description: 新增字段
        type: string
      children:
        items:
          $ref: '#/definitions/normaladmin_backend_internal_models.Menu'
        type: array
      component:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_hidden:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
      parent_name:
        type: string
      path:
        type: string
      permission:
        type: string
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        description: menu or button
        type: string
    type: object
  normaladmin_backend_internal_models.Role:
    properties:
      code:
        type: string
      data_scope:
        description: 数据权限范围
        type: integer
      description:
        type: string
      id:
        description: 将 ID 字段的 JSON 标签设置为 "id"
        type: integer
      is_preset:
        description: 是否为系统预设角色
        type: boolean
      name:
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
    type: object
  normaladmin_backend_pkg_utils_response.ResponseData:
    properties:
      data: {}
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /gam/roles:
    get:
      consumes:
      - application/json
      description: 获取角色列表，支持按名称搜索、分页和排序
      parameters:
      - description: 角色名称，用于搜索过滤
        in: query
        name: name
        type: string
      - default: 1
        description: 页码，默认值：1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认值：10
        in: query
        name: pageSize
        type: integer
      - description: 排序字段，多个字段用逗号分隔
        example: '"id,created_at"'
        in: query
        name: sortField
        type: string
      - description: 排序方式，多个排序方式用逗号分隔(asc/desc)
        example: '"desc,asc"'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    roles:
                      items:
                        $ref: '#/definitions/normaladmin_backend_internal_models.Role'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 获取角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建新的角色，包括角色名称、描述等信息
      parameters:
      - description: 角色信息，包括名称、描述等
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/normaladmin_backend_internal_models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    role:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Role'
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 创建角色
      tags:
      - 角色管理
  /gam/roles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据角色ID删除指定角色
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的角色ID
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 根据角色ID获取角色的详细信息
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    role:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Role'
                  type: object
              type: object
        "400":
          description: 无效的角色ID
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 获取单个角色
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 根据角色ID更新角色信息
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 需要更新的角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/normaladmin_backend_internal_models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    role:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Role'
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 更新角色
      tags:
      - 角色管理
  /gam/roles/{id}/sort:
    put:
      consumes:
      - application/json
      description: 更新指定角色的排序值
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 排序信息
        in: body
        name: sort
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 更新角色排序
      tags:
      - 角色管理
  /gam/roles/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新指定角色的启用/禁用状态
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 更新角色状态
      tags:
      - 角色管理
  /gam/roles/{roleId}/menus:
    get:
      consumes:
      - application/json
      description: 获取指定角色的菜单权限树和已选中的菜单列表
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    checkedMenus:
                      items:
                        type: integer
                      type: array
                    menuTree:
                      items:
                        $ref: '#/definitions/normaladmin_backend_internal_models.Menu'
                      type: array
                  type: object
              type: object
        "400":
          description: 无效的角色ID
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 获取角色菜单权限
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 更新指定角色的菜单权限
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: roleId
        required: true
        type: integer
      - description: 菜单ID列表
        in: body
        name: menuIds
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 更新角色菜单权限
      tags:
      - 角色管理
  /gam/roles/check-field:
    get:
      consumes:
      - application/json
      description: 检查指定字段的值在角色表中是否唯一
      parameters:
      - description: 字段名称
        example: '"name"'
        in: query
        name: field
        required: true
        type: string
      - description: 字段值
        example: '"admin"'
        in: query
        name: value
        required: true
        type: string
      - description: 需要排除的角色ID
        in: query
        minimum: 1
        name: excludeId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    unique:
                      type: boolean
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 检查角色字段唯一性
      tags:
      - 角色管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"